refactor: Video controller 패키지 리펙토링 및 @AutoConfigureWebTestClient 추가
