+++ src/test/java/com/wootube/ioi/web/config/TestConfig.java
@@ -1,4 +1,4 @@
-package com.wootube.ioi.web.controller;
+package com.wootube.ioi.web.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.AnonymousAWSCredentials;
